requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
build==0.10.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0
python-http-client==3.3.7
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
sendgrid==6.10.0
soupsieve==2.4.1
starkbank-ecdsa==2.2.0
urllib3==2.0.4
waitress==2.1.2
Werkzeug==2.3.6
WTForms==3.0.1
